+++ src/Genius/Resources/AbstractResource.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace Genius\Resources;
 
-use Genius\Authentication\Scope;
+use Genius\Authentication\OAuth2;
 use Genius\Genius;
 
 class AbstractResource
@@ -37,12 +37,28 @@ class AbstractResource
      */
     protected function requireScope(string $scope): bool
     {
-        /** @var Scope $scopes */
+        if (!($this->genius->getAuthentication() instanceof OAuth2)) {
+            throw new ResourceException(sprintf(
+                '"%s" requires "%s" scope which is available only when using OAuth2 authentication.',
+                $this->getCallerClassAndFunctionName(),
+                $scope
+            ));
+        }
+
         $scopes = $this->genius->getAuthentication()->getScope();
         if ($scopes->hasScope($scope)) {
             return true;
         }
         
-        throw new ResourceException(sprintf('You have no access to required scope %s for that action.', $scope));
+        throw new ResourceException(sprintf(
+            'You have no access to required scope "%s" for "%s" action.',
+            $scope,
+            $this->getCallerClassAndFunctionName()
+        ));
+    }
+
+    private function getCallerClassAndFunctionName(): string
+    {
+        return debug_backtrace()[2]['class'] . '::' . debug_backtrace()[2]['function'];
     }
 }