+++ src/Genius/Resources/ResourceException.php
@@ -3,7 +3,9 @@ declare(strict_types=1);
 
 namespace Genius\Resources;
 
-class ResourceException extends \Exception
+use Genius\GeniusException;
+
+final class ResourceException extends GeniusException
 {
     
 }